Added stopwatch used to maintain delta-time

+++ Millie-Quest-for-Dinner/Camera.cs
@@ -8,8 +8,8 @@ namespace Millie_Quest_for_Dinner
 {
     public class Camera
     {
-        private int _x;
-        private int _y;
+        private double _x;
+        private double _y;
 
         /* 
         private int _width;
